+++ package.json
@@ -77,6 +77,7 @@
     "react-router": "^3.0.0",
     "shx": "^0.2.1",
     "superagent": "^3.3.0",
+    "url": "^0.11.0",
     "webpack": "^2.2.1"
   },
   "engines": {
